fea: move event listeners registration
- separate module and service file
