+++ src/events/mod.ts
@@ -1,3 +1,3 @@
-export * from './constants.ts';
 export * from './decorator.ts';
 export * from './events.ts';
+export * from './module.ts';
